Add files via upload
This is the second PowerBI project I'm taking, thanks to the ongoing DSN Unilorin Datathon. This visualisation shows the type of meats and quantity produced in Germany from 2000 to 2022. The dataset was gotten from Dataworks. The three main types of meat for this dataset includes Poultry, Pork and Beef. Check file for more details.